User LOgged in or not

+++ resources/views/list.blade.php
@@ -37,8 +37,10 @@
       <td>{{$item->company_name}}</td>
       <td>{{$item->joining_month}}</td>
       <td>{{$item->profile}}</td>
+@if(Session::get('email') == {{$item->email}})
       <td><a href="delete/{{$item->email}}"><i class="fa fa-trash"></i></a></td>
       <td><a href="edit/{{$item->email}}"><i class="fa fa-edit"></i></a></td>
+@endif
     </tr>
     @endforeach
   </tbody>
